refactor(productivity): implement ViewModel pattern and extract UI components
Decompose the massive productivity.tsx screen into smaller components: ProductivityHeader, ProductivityTabs, ProductivityEmptyState, and TodayPlanBanner. Move all state, data manipulation, and filtering logic into a dedicated useProductivityViewModel hook. Implement dynamic theming via useTheme.
